Create Roster
Roster class that has internal methods to create an arraylist from a comma separated array; go through the arraylist for correct email addresses, print out the entire arraylist, print only some of the arraylist, remove a record of the arraylist, attempt to remove the same record and print out a message if the record doesn't exist. Go through the array and get the grades for each student and spit out the avg of those grades.